Applied fixes and Organized files within the components folder
