+++ src/App.js
@@ -1,9 +1,8 @@
-import React, { useState } from "react";
-
-import "./App.css";
-import NotFound from "./NotFound";
-import InputForm from "./InputForm";
-import UserProfile from "./UserProfile";
+import React, { useState } from 'react';
+import InputForm from './components/InputForm';
+import UserProfile from './components/UserProfile';
+import NotFound from './components/NotFound';
+import './App.css';
 
 const App = () => {
   const [user, setUser] = useState(null);
@@ -12,6 +11,7 @@ const App = () => {
   const [showHeading, setShowHeading] = useState(true);
   const [showParagraph, setShowParagraph] = useState(true); // State to control visibility of the paragraph
   const [hireClicked, setHireClicked] = useState(false); // State to track if "Hire Me" button is clicked
+  
 
   const fetchUserData = async (username) => {
     try {
@@ -23,7 +23,7 @@ const App = () => {
         setShowHeading(false);
         setShowParagraph(false);
       } else {
-        throw new Error("User not found");
+        throw new Error('User not found');
       }
     } catch (error) {
       setError(error.message);
@@ -39,21 +39,18 @@ const App = () => {
   return (
     <div>
       {showHeading && <h1>GitHub Profile</h1>}
-      {showParagraph && <p>Generate your GitHub Profile</p>}{" "}
+      {showParagraph && <p>Generate your GitHub Profile</p>} {/* Conditionally render the paragraph */}
       <InputForm onSubmit={fetchUserData} visible={showInputForm} />
       {error ? (
         <NotFound />
       ) : user ? (
         <React.Fragment>
           <UserProfile user={user} />
-          <button className="hire-button" onClick={handleHireClick}>
-            Hire Me
-          </button>
-          {hireClicked && <p>Hire Me button clicked!</p>}{" "}
+          <button className="hire-button" onClick={handleHireClick}>Hire Me</button>
+          {hireClicked && <p>Hire Me button clicked!</p>} {/* Display message when the button is clicked */}
         </React.Fragment>
       ) : null}
     </div>
   );
 };
-
 export default App;
